refactor(engine): rename ImmutableZeebeState to ZeebeState
