+++ engine/src/main/java/io/zeebe/engine/state/ZeebeState.java
@@ -5,11 +5,23 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.zeebe.engine.state.immutable;
+package io.zeebe.engine.state;
 
-import io.zeebe.engine.state.ZbColumnFamilies;
+import io.zeebe.engine.state.immutable.BlackListState;
+import io.zeebe.engine.state.immutable.DeploymentState;
+import io.zeebe.engine.state.immutable.ElementInstanceState;
+import io.zeebe.engine.state.immutable.EventScopeInstanceState;
+import io.zeebe.engine.state.immutable.IncidentState;
+import io.zeebe.engine.state.immutable.JobState;
+import io.zeebe.engine.state.immutable.MessageStartEventSubscriptionState;
+import io.zeebe.engine.state.immutable.MessageState;
+import io.zeebe.engine.state.immutable.MessageSubscriptionState;
+import io.zeebe.engine.state.immutable.TimerInstanceState;
+import io.zeebe.engine.state.immutable.VariableState;
+import io.zeebe.engine.state.immutable.WorkflowInstanceSubscriptionState;
+import io.zeebe.engine.state.immutable.WorkflowState;
 
-public interface ImmutableZeebeState {
+public interface ZeebeState {
 
   DeploymentState getDeploymentState();
